# specs/001-aeo-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Client AEO Platform API
  version: 1.0.0
  description: |
    API for managing AI-powered Answer Engine Optimization crawling, analysis, and scoring across multiple organizations and projects.

    Features include:
    - Multi-tenant organization management with role-based access
    - Project-based crawl configuration
    - Intelligent page-type-aware scoring (0-100 scale)
    - AI-powered recommendations
    - Historical tracking and comparison
    - Token usage monitoring
  contact:
    name: AEO Platform Support
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://your-vercel-app.vercel.app/api
    description: Production server
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Create new user account
      description: Register a new user account with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                name:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                  name:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate and get JWT token
      description: Login with email and password to receive authentication token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT bearer token
                  expiresIn:
                    type: integer
                    description: Token expiration in seconds
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                      name:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                  name:
                    type: string
                  organizations:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                        role:
                          type: string
                          enum: [admin, editor, viewer]
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /projects/{projectId}/crawls:
    parameters:
      - $ref: '#/components/parameters/projectId'
    post:
      tags:
        - Crawl Runs
      summary: Start new crawl
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                runType:
                  type: string
                  enum: [full, sitemap_only, sample, delta]
                  default: full
      responses:
        '201':
          description: Crawl started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlRun'

    get:
      tags:
        - Crawl Runs
      summary: List crawl runs
      responses:
        '200':
          description: Crawl runs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CrawlRun'
  /crawls/{crawlId}:
    parameters:
      - $ref: '#/components/parameters/crawlId'
    get:
      tags:
        - Crawl Runs
      summary: Get crawl status
      responses:
        '200':
          description: Crawl status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlRun'
  /crawls/{crawlId}/pause:
    parameters:
      - $ref: '#/components/parameters/crawlId'
    post:
      tags:
        - Crawl Runs
      summary: Pause crawl
      responses:
        '200':
          description: Crawl paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlRun'
  /crawls/{crawlId}/resume:
    parameters:
      - $ref: '#/components/parameters/crawlId'
    post:
      tags:
        - Crawl Runs
      summary: Resume crawl
      responses:
        '200':
          description: Crawl resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlRun'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    crawlId:
      name: crawlId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CrawlRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        runType:
          type: string
          enum: [full, sitemap_only, sample, delta]
        status:
          type: string
          enum: [queued, running, paused, completed, failed]
        pagesDiscovered:
          type: integer
        pagesProcessed:
          type: integer
        tokenUsage:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...